Rename file with patch
