+++ kubemarine/patches/p3_managed_sandbox_image.py
